Mark the footprint as virtual in the pretty format
The export currently adds (attr smd), which marks the footprint as an
SMD component (which internally sets the MOD_CMS attribute, and in the
GUI marks the component as "Normal+Insert").

This causes it to be exported in a .pos file for a pick & place machine.
Since this is just a silkscreen and not an actual component, this makes
no sense.

This commit instead sets (attr virtual) (which internally sets
MOD_VIRTUAL, and in the GUI marks the component as "Virtual") which
causes it to be ignored by various parts of kicad that iterate over
actual components.