Write the full svg2mod commandline to the output file
Previously, just the input filename was documented in the output file,
but it is also useful to document the options that were used (i.e.
precision, scaling, etc). This makes it easier to later redo the
conversion after changing the input file, or making another size of the
same file, etc.

The easiest way to document all relevant options is to just write down
the full commandline used, which is what this commit does. This is put
into the description field for pretty files and the comment header for
legacy files, replacing the input filename that was already there.

Note that the legacy updater outupt code is not changed, since when
updating a legacy footprint file rather than creating it, the header is
copied from the original file rather than writing a new one. This could
cause the command in the header to not actually match the content of the
file. However, because this problem already exists for the filename in
the header, and the legacy format is probably not actively used much if
at all, this seems acceptable to me.