Apply commandline scaling factor to stroke widths
Previously, using the --factor option would scale all points, but keep
stroke widths at the same size, which could make the stroke widths way
too small or large, especially with big scaling.

With this commit the Svg2ModExport.scale_factor is used to calculate the
stroke width, which includes the commandline scaling factor. Since it
also includes the dpi conversion and (when needed), the decimil
conversion, those conversions can be removed for the stroke width (so
this changes the result of _get_fill_stroke to be output units (mm or
decimil), rather than always mm.

This fixes a part of #25.